A2 - 91%
This was the second assignment for this module.
We had to create a table view to display a list of artworks on the University of Liverpool campus, organised by proximity to the user's location. When one of the artworks in the table is clicked on, the user should be shown a new view with more detailed information about that piece of artwork, including an image.
Each building containing a piece of artwork also had to have a pin on the map view. When the pin is clicked on a list of artworks in that building should be displayed (if there is more than one) and the user can then click on any of them to read more information, or the user will directly be taken to the 'more information' view if it is the only artwork in that building.